Ray tracer

+++ src/camera.cpp
@@ -3,10 +3,10 @@
 camera::camera(point3 lookfrom,
 		point3 lookat,
 		vec3 vup,
-		double vfov,
-		double aspect_ratio,
-		double aperture,
-		double focus_dist)
+		FLOAT vfov,
+		FLOAT aspect_ratio,
+		FLOAT aperture,
+		FLOAT focus_dist)
 {
 	auto theta = degrees_to_radians(vfov);
 	auto h = tan(theta/2);
@@ -25,7 +25,7 @@ camera::camera(point3 lookfrom,
 	lens_radius = aperture / 2;
 }
 
-ray camera::get_ray(double s, double t) const
+ray camera::get_ray(FLOAT s, FLOAT t) const
 {
 	vec3 rd = lens_radius * random_in_unit_disk();
 	vec3 offset = u * rd.x() + v * rd.y();
